modularize UserSettingsInterest
